Clean up more includes
